freqBarDef renamed to freqBar. Successfully generates the diagram.
