+++ health/src/main/java/com/health/visuals/FreqBar.java
@@ -1,153 +1,73 @@
 package com.health.visuals;
 
-import static com.googlecode.charts4j.Color.*;
-
 import java.util.ArrayList;
-import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
 
-import com.googlecode.charts4j.AxisLabels;
-import com.googlecode.charts4j.AxisLabelsFactory;
-import com.googlecode.charts4j.AxisStyle;
-import com.googlecode.charts4j.AxisTextAlignment;
-import com.googlecode.charts4j.BarChart;
-import com.googlecode.charts4j.BarChartPlot;
-import com.googlecode.charts4j.Data;
-import com.googlecode.charts4j.DataUtil;
-import com.googlecode.charts4j.Fills;
-import com.googlecode.charts4j.GCharts;
-import com.googlecode.charts4j.Plots;
-
-import java.io.IOException;
-import java.net.URL;
-
+import com.health.Chunk;
+import com.health.Column;
+import com.health.Record;
+import com.health.Table;
+import com.xeiam.xchart.Chart;
+import com.xeiam.xchart.ChartBuilder;
+import com.xeiam.xchart.StyleManager.ChartType;
+import com.xeiam.xchart.StyleManager.LegendPosition;
+import com.xeiam.xchart.SwingWrapper;
 
 /**
- * Creates Bar charts.
- * @author lizzy
+ * Generates a Frequency Bar Diagram based on a Table object.
+ * @author Bjorn van der Laan & Lizzy Scholten
  *
  */
-public final class FreqBar {
+public class FreqBar {
+    /**
+     * Generates a Frequency Bar diagram.
+     * 
+     * @param table
+     *            Table to use
+     * @param column
+     *            Column to display frequency of
+     */
+    public static void frequencyBar(final Table table, final String column) {
+        // Get the Column to calculate frequency of
+        Column selectedColumn = table.getColumn(column);
+        String columnName = selectedColumn.getName();
 
-	private FreqBar() { }
+        // Create map to save frequencies
+        Map<String, Integer> freqMap = new HashMap<String, Integer>();
 
-	/**
-	 * Possible preprocessing of info.
-	 * @param arIn			the data
-	 * @param title			title of chart
-	 * @return 				the url of the image
-	 * @throws IOException	if input/output exception of some sort has occurred
-	 */
-	public static URL makeBarChart(final double[] arIn, final String title) throws IOException {
-		double[][] arOut = new double[arIn.length][2];
-		// For checking if event is already
-		boolean check = false;
+        for (Chunk c : table) {
+            for (Record r : c) {
+                // Get value of record
+                String value = r.getValue(columnName).toString();
+                if (!freqMap.containsKey(value)) {
+                    freqMap.put(value, 1);
+                } else {
+                    int currentFrequency = freqMap.get(value);
+                    freqMap.replace(value, ++currentFrequency);
+                }
+            }
+        }
+        makeBarChart(freqMap);
+    }
 
-		int count = 0;
-		for (int i = 0; i < arIn.length; i++) {
-			// Check if value is already present in arOut
-			for (int k = 0; k < arIn.length; k++) {
-				if (arOut[k][0] == arIn[i]) {
-					int ind = k;
-					double val = arOut[ind][1];
-					val = val + 1;
-					arOut[ind][1] = val;
-					check = true;
-					break;
-				}
-			}
-			// Else put it in arOut
-			if (!check) {
-				arOut[count][0] = arIn[i];
-				arOut[count][1] = 1;
-				count = count + 1;
-			}
-			check = false;
-		}
+    private static void makeBarChart(final Map<String, Integer> freqMap) {
+        // Convert input data for processing
+        ArrayList<String> labels = new ArrayList<String>(freqMap.keySet());
+        ArrayList<Integer> frequency = new ArrayList<Integer>(freqMap.values());
 
-		URL url = makeBarChart(arOut, title);
-		return url;
-	}
+        // Create Chart
+        Chart chart = new ChartBuilder().chartType(ChartType.Bar).width(800)
+                .height(600).title("Score Histogram").xAxisTitle("Event")
+                .yAxisTitle("Frequency").build();
 
-	/**
-	 * Creates url for Bar Chart.
-	 * @param arIn			the data
-	 * @param title			title of chart
-	 * @return 				the url of the image
-	 * @throws IOException	if input/output exception of some sort has occurred
-	 */
-	public static URL makeBarChart(final double[][] arIn, final String title) throws IOException {
-		// Defining data
-		double[] dataAr = new double[arIn.length];
-		// Set count to 0
-		int count = 0;
-		// Copy second column of input array into dataArray
-		for (int i = 0; i < arIn.length; i++) {
-			dataAr[count] = arIn[i][1];
-			count = count + 1;
-		}
-		// Get maximum value of dataArray
-		double maxVal = dataAr[0];
-		for (int m = 0; m < dataAr.length; m++) {
-			if (dataAr[m] > maxVal) {
-				maxVal = dataAr[m];
-			}
-		}
-		// Defining labels
-		String[] labelsAr = new String[count];
-		count = 0;
-		// Create labels
-		for (int k = 0; k < arIn.length; k++) {
-			labelsAr[count] = String.valueOf(arIn[k][0]);
-			count = count + 1;
-		}
-		// Turn labels into list
-		ArrayList<String> list = new ArrayList<String>(Arrays.asList(labelsAr));
-		ArrayList<String> newList = new ArrayList<String>();
-		// Delete values from original label array that are useless and that are not going to be used as labels
-		// Do this by copying useful ones in new ArrayList
-		for (int n = 0; n < list.size(); n++) {
-			if (!list.get(n).equals("0.0") && !list.get(n).equals(null)) {
-				newList.add(list.get(n));
-			}
-		}
+        chart.addSeries("Test 1", new ArrayList<String>(labels),
+                new ArrayList<Integer>(frequency));
 
-		// Transform labels list to array
-		String[] labelsUse = newList.toArray(new String[newList.size()]);
+        // Customize Chart
+        chart.getStyleManager().setLegendPosition(LegendPosition.InsideNW);
 
-		// Add data to the chart
-		final double maxEvent = maxVal + 1;
-		Data data = DataUtil.scaleWithinRange(0,  maxEvent, dataAr);
-		BarChartPlot dat = Plots.newBarChartPlot(data, RED, "Data");
-		BarChart chart = GCharts.newBarChart(dat);
+        new SwingWrapper(chart).displayChart();
+    }
 
-		//Defining axis (labels etc)
-		AxisStyle axisStyle = AxisStyle.newAxisStyle(BLACK, 13, AxisTextAlignment.CENTER);
-		AxisLabels event = AxisLabelsFactory.newAxisLabels("Event", 50.0);
-		event.setAxisStyle(axisStyle);
-		AxisLabels events = AxisLabelsFactory.newAxisLabels(Arrays.asList(labelsUse));
-		events.setAxisStyle(axisStyle);
-		AxisLabels frequency = AxisLabelsFactory.newAxisLabels("Frequency", 50.0);
-		frequency.setAxisStyle(axisStyle);
-		AxisLabels freqCount = AxisLabelsFactory.newNumericRangeAxisLabels(0, maxEvent);
-		freqCount.setAxisStyle(axisStyle);
-
-		// Add the axis info to the chart.
-        chart.addYAxisLabels(freqCount);
-        chart.addYAxisLabels(frequency);
-        chart.addXAxisLabels(events);
-        chart.addXAxisLabels(event);
-        chart.setHorizontal(false);
-
-        //size should be calculate based on frequencies and amount of labels
-        chart.setSize(650, 450);
-        chart.setSpaceBetweenGroupsOfBars(20);
-
-        // Set title and background
-        chart.setTitle(title, BLACK, 17);
-        chart.setBackgroundFill(Fills.newSolidFill(LIGHTGREY));
-        // Create url of the chart
-        URL url = new URL(chart.toURLString());
-
-        return (url);
-	}
 }
