+++ health/src/main/java/com/health/visuals/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author Bjorn van der Laan
+ *
+ */
+package com.health.visuals;
